Create Game, Team, and Venue tables in the database
Also index the foreign keys in the entities

+++ app/src/main/java/com/c50x/eleos/data/GameDao.java
@@ -15,10 +15,10 @@ public interface GameDao {
     Long addGame(Game game);
 
     @Update
-    Long modifyGame(Game game);
+    int modifyGame(Game game);
 
     @Delete
-    Long deleteGame(Game game);
+    int deleteGame(Game game);
 
     @Query("SELECT * FROM Game WHERE gameId LIKE:id")
     Game[] loadGame(int id);
